remove JSDoc, fix typings, rewrite tests with native node test suite (wip)
