+++ .github/workflows/ci.yml
@@ -14,10 +14,10 @@ jobs:
       - uses: actions/checkout@v3
       - uses: actions/setup-node@v3
         with:
-          node-version: 18
+          node-version: 20
           cache: 'npm'
           cache-dependency-path: '**/package-lock.json'
       - run: npm ci --ignore-scripts
       - run: npm run lint
-      - run: npm run test:ci
+      - run: npm run test
       - run: npm run build
