+++ package.json
@@ -12,15 +12,16 @@
   "bugs": {
     "url": "https://github.com/placekit/client-js/issues"
   },
+  "type": "module",
   "types": "./dist/placekit.d.ts",
-  "module": "./dist/placekit.esm.mjs",
-  "main": "./dist/placekit.cjs.js",
+  "module": "./dist/placekit.js",
+  "main": "./dist/placekit.cjs",
   "browser": "./dist/placekit.umd.js",
   "exports": {
     ".": {
       "types": "./dist/placekit.d.ts",
-      "require": "./dist/placekit.cjs.js",
-      "import": "./dist/placekit.esm.mjs"
+      "require": "./dist/placekit.cjs",
+      "import": "./dist/placekit.js"
     }
   },
   "files": [
@@ -32,18 +33,13 @@
     "build": "rollup -c",
     "lint": "eslint ./src",
     "format": "eslint ./src --fix",
-    "test": "jest --runInBand --detectOpenHandles --forceExit --roots ./src",
-    "test:watch": "jest --runInBand --detectOpenHandles --forceExit --roots ./src --onlyFailures --watch",
-    "test:ci": "jest --runInBand --detectOpenHandles --forceExit --roots ./src --ci",
-    "install:examples": "find examples -maxdepth 2 -name package.json -not -path \"*/node_modules/*\" -execdir npm i \\;"
+    "test": "node --test ./src",
+    "test:watch": "node --test ./src --watch"
   },
   "devDependencies": {
     "@rollup/plugin-commonjs": "^25.0.2",
     "eslint": "^8.43.0",
     "eslint-config-google": "^0.14.0",
-    "jest": "^29.5.0",
-    "jest-environment-jsdom": "^29.5.0",
-    "jest-fetch-mock": "^3.0.3",
     "rimraf": "^5.0.1",
     "rollup": "^3.25.2",
     "rollup-plugin-cleanup": "^3.2.1",
